optimize structure for table database

+++ includes/sqldata.php
@@ -37,108 +37,108 @@ function nv_create_table_sys( $lang )
 
 	$sql_create_table = array();
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_modules (
-	 title varchar(55) NOT NULL,
-	 module_file varchar(55) NOT NULL DEFAULT '',
-	 module_data varchar(55) NOT NULL DEFAULT '',
-	 custom_title varchar(255) NOT NULL,
-	 admin_title varchar(255) NOT NULL,
-	 set_time int(11) unsigned NOT NULL DEFAULT '0',
-	 main_file tinyint(1) unsigned NOT NULL DEFAULT '0',
-	 admin_file tinyint(1) unsigned NOT NULL DEFAULT '0',
-	 theme varchar(100) NOT NULL,
-	 mobile varchar(100) NOT NULL,
- 	 description varchar(255) NOT NULL DEFAULT '',
-	 keywords mediumtext NOT NULL,
-	 groups_view varchar(255) NOT NULL,
-	 in_menu tinyint(1) unsigned NOT NULL DEFAULT '0',
-	 weight tinyint(3) unsigned NOT NULL DEFAULT '1',
-	 submenu tinyint(1) unsigned NOT NULL DEFAULT '0',
-	 act tinyint(1) unsigned NOT NULL DEFAULT '0',
-	 admins varchar(255) NOT NULL,
-	 rss tinyint(4) NOT NULL default '1',
-	 gid MEDIUMINT(8) NOT NULL default '0',
-	 PRIMARY KEY (title)
+		 title varchar(55) NOT NULL,
+		 module_file varchar(55) NOT NULL DEFAULT '',
+		 module_data varchar(55) NOT NULL DEFAULT '',
+		 custom_title varchar(255) NOT NULL,
+		 admin_title varchar(255) NOT NULL,
+		 set_time int(11) unsigned NOT NULL DEFAULT '0',
+		 main_file tinyint(1) unsigned NOT NULL DEFAULT '0',
+		 admin_file tinyint(1) unsigned NOT NULL DEFAULT '0',
+		 theme varchar(100) NOT NULL,
+		 mobile varchar(100) NOT NULL,
+	 	 description varchar(255) NOT NULL DEFAULT '',
+		 keywords mediumtext NOT NULL,
+		 groups_view varchar(255) NOT NULL,
+		 in_menu tinyint(1) unsigned NOT NULL DEFAULT '0',
+		 weight tinyint(3) unsigned NOT NULL DEFAULT '1',
+		 submenu tinyint(1) unsigned NOT NULL DEFAULT '0',
+		 act tinyint(1) unsigned NOT NULL DEFAULT '0',
+		 admins varchar(255) NOT NULL,
+		 rss tinyint(4) NOT NULL default '1',
+		 gid smallint(5) NOT NULL default '0',
+		 PRIMARY KEY (title)
 	) ENGINE=MyISAM";
 
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_blocks_groups (
-	 bid int(11) unsigned NOT NULL AUTO_INCREMENT,
-	 theme varchar(55) NOT NULL,
-	 module varchar(55) NOT NULL,
-	 file_name varchar(55) DEFAULT NULL,
-	 title varchar(255) DEFAULT NULL,
-	 link varchar(255) DEFAULT NULL,
-	 template varchar(55) DEFAULT NULL,
-	 position varchar(55) DEFAULT NULL,
-	 exp_time int(11) DEFAULT '0',
-	 active tinyint(4) DEFAULT '0',
-	 groups_view varchar(255) DEFAULT '',
-	 all_func tinyint(4) NOT NULL DEFAULT '0',
-	 weight int(11) NOT NULL DEFAULT '0',
-	 config text,
-	 PRIMARY KEY (bid),
-	 KEY theme (theme),
-	 KEY module (module),
-	 KEY position (position),
-	 KEY exp_time (exp_time)
+		 bid mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
+		 theme varchar(55) NOT NULL,
+		 module varchar(55) NOT NULL,
+		 file_name varchar(55) DEFAULT NULL,
+		 title varchar(255) DEFAULT NULL,
+		 link varchar(255) DEFAULT NULL,
+		 template varchar(55) DEFAULT NULL,
+		 position varchar(55) DEFAULT NULL,
+		 exp_time int(11) DEFAULT '0',
+		 active tinyint(4) DEFAULT '0',
+		 groups_view varchar(255) DEFAULT '',
+		 all_func tinyint(4) NOT NULL DEFAULT '0',
+		 weight int(11) NOT NULL DEFAULT '0',
+		 config text,
+		 PRIMARY KEY (bid),
+		 KEY theme (theme),
+		 KEY module (module),
+		 KEY position (position),
+		 KEY exp_time (exp_time)
 	) ENGINE=MyISAM";
 
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_blocks_weight (
-	 bid int(11) NOT NULL DEFAULT '0',
-	 func_id int(11) NOT NULL DEFAULT '0',
-	 weight int(11) NOT NULL DEFAULT '0',
-	 UNIQUE KEY bid (bid,func_id)
+		 bid mediumint(8) NOT NULL DEFAULT '0',
+		 func_id mediumint(8) NOT NULL DEFAULT '0',
+		 weight mediumint(8) NOT NULL DEFAULT '0',
+		 UNIQUE KEY bid (bid,func_id)
 	) ENGINE=MyISAM";
 
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_modfuncs (
-	 func_id mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
-	 func_name varchar(55) NOT NULL,
-	 alias varchar(55) NOT NULL DEFAULT '',
-	 func_custom_name varchar(255) NOT NULL,
-	 in_module varchar(55) NOT NULL,
-	 show_func tinyint(4) NOT NULL DEFAULT '0',
-	 in_submenu tinyint(1) unsigned NOT NULL DEFAULT '0',
-	 subweight smallint(2) unsigned NOT NULL DEFAULT '1',
-	 setting varchar(255) NOT NULL DEFAULT '',
-	 PRIMARY KEY (func_id),
-	 UNIQUE KEY func_name (func_name,in_module),
-	 UNIQUE KEY alias (alias,in_module)
+		 func_id mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
+		 func_name varchar(55) NOT NULL,
+		 alias varchar(55) NOT NULL DEFAULT '',
+		 func_custom_name varchar(255) NOT NULL,
+		 in_module varchar(55) NOT NULL,
+		 show_func tinyint(4) NOT NULL DEFAULT '0',
+		 in_submenu tinyint(1) unsigned NOT NULL DEFAULT '0',
+		 subweight smallint(2) unsigned NOT NULL DEFAULT '1',
+		 setting varchar(255) NOT NULL DEFAULT '',
+		 PRIMARY KEY (func_id),
+		 UNIQUE KEY func_name (func_name,in_module),
+		 UNIQUE KEY alias (alias,in_module)
 	) ENGINE=MyISAM";
 
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_searchkeys (
-	 id varchar(32) NOT NULL DEFAULT '',
-	 skey varchar(255) NOT NULL,
-	 total int(11) NOT NULL DEFAULT '0',
-	 search_engine varchar(50) NOT NULL,
-	 KEY (id),
-	 KEY skey (skey),
-	 KEY search_engine (search_engine)
+		 id varchar(32) NOT NULL DEFAULT '',
+		 skey varchar(255) NOT NULL,
+		 total int(11) NOT NULL DEFAULT '0',
+		 search_engine varchar(50) NOT NULL,
+		 KEY (id),
+		 KEY skey (skey),
+		 KEY search_engine (search_engine)
 	) ENGINE=MyISAM";
 
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_referer_stats (
-	 host varchar(255) NOT NULL,
-	 total int(11) NOT NULL DEFAULT '0',
-	 month01 int(11) NOT NULL DEFAULT '0',
-	 month02 int(11) NOT NULL DEFAULT '0',
-	 month03 int(11) NOT NULL DEFAULT '0',
-	 month04 int(11) NOT NULL DEFAULT '0',
-	 month05 int(11) NOT NULL DEFAULT '0',
-	 month06 int(11) NOT NULL DEFAULT '0',
-	 month07 int(11) NOT NULL DEFAULT '0',
-	 month08 int(11) NOT NULL DEFAULT '0',
-	 month09 int(11) NOT NULL DEFAULT '0',
-	 month10 int(11) NOT NULL DEFAULT '0',
-	 month11 int(11) NOT NULL DEFAULT '0',
-	 month12 int(11) NOT NULL DEFAULT '0',
-	 last_update int(11) NOT NULL DEFAULT '0',
-	 UNIQUE KEY host (host),
-	 KEY total (total)
+		 host varchar(255) NOT NULL,
+		 total int(11) NOT NULL DEFAULT '0',
+		 month01 int(11) NOT NULL DEFAULT '0',
+		 month02 int(11) NOT NULL DEFAULT '0',
+		 month03 int(11) NOT NULL DEFAULT '0',
+		 month04 int(11) NOT NULL DEFAULT '0',
+		 month05 int(11) NOT NULL DEFAULT '0',
+		 month06 int(11) NOT NULL DEFAULT '0',
+		 month07 int(11) NOT NULL DEFAULT '0',
+		 month08 int(11) NOT NULL DEFAULT '0',
+		 month09 int(11) NOT NULL DEFAULT '0',
+		 month10 int(11) NOT NULL DEFAULT '0',
+		 month11 int(11) NOT NULL DEFAULT '0',
+		 month12 int(11) NOT NULL DEFAULT '0',
+		 last_update int(11) NOT NULL DEFAULT '0',
+		 UNIQUE KEY host (host),
+		 KEY total (total)
 	) ENGINE=MyISAM";
 
 	$sql_create_table[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_modthemes (
-	 func_id int(11) DEFAULT NULL,
-	 layout varchar(100) DEFAULT NULL,
-	 theme varchar(100) DEFAULT NULL,
-	 UNIQUE KEY func_id (func_id,layout,theme)
+		 func_id mediumint(8) DEFAULT NULL,
+		 layout varchar(100) DEFAULT NULL,
+		 theme varchar(100) DEFAULT NULL,
+		 UNIQUE KEY func_id (func_id,layout,theme)
 	 ) ENGINE=MyISAM";
 
  	$sql_create_table[] = "INSERT INTO " . $db_config['prefix'] . "_" . $lang . "_modules (title, module_file, module_data, custom_title, admin_title, set_time, main_file, admin_file, theme, mobile, description, keywords, groups_view, in_menu, weight, submenu, act, admins, rss, gid) VALUES
@@ -155,20 +155,20 @@ function nv_create_table_sys( $lang )
 		('page', 'page', 'page', 'page', '', 1279360267, 1, 1, '', '', '', '', '0', 0, 11, 1, 1, '', 0, 0)";
 
 	$sql_create_table[] = "INSERT INTO " . NV_CONFIG_GLOBALTABLE . " (lang, module, config_name, config_value) VALUES
- ('" . $lang . "', 'global', 'site_name', 'NukeViet CMS 3.x'),
- 	('" . $lang . "', 'global', 'site_logo', 'images/logo.png'),
- 	('" . $lang . "', 'global', 'site_description', 'NukeViet CMS 3.x Developed by VINADES.,JSC'),
- 	('" . $lang . "', 'global', 'site_keywords', ''),
- ('" . $lang . "', 'global', 'site_theme', '" . $global_config['site_theme'] . "'),
- ('" . $lang . "', 'global', 'site_home_module', 'users'),
- ('" . $lang . "', 'global', 'switch_mobi_des', '1'),
+	 	('" . $lang . "', 'global', 'site_name', 'NukeViet CMS 3.x'),
+	 	('" . $lang . "', 'global', 'site_logo', 'images/logo.png'),
+	 	('" . $lang . "', 'global', 'site_description', 'NukeViet CMS 3.x Developed by VINADES.,JSC'),
+	 	('" . $lang . "', 'global', 'site_keywords', ''),
+	 	('" . $lang . "', 'global', 'site_theme', '" . $global_config['site_theme'] . "'),
+	 	('" . $lang . "', 'global', 'site_home_module', 'users'),
+	 	('" . $lang . "', 'global', 'switch_mobi_des', '1'),
 		('" . $lang . "', 'global', 'upload_logo', 'images/logo.png'),
 		('" . $lang . "', 'global', 'autologosize1', '50'),
 		('" . $lang . "', 'global', 'autologosize2', '40'),
 		('" . $lang . "', 'global', 'autologosize3', '30'),
 		('" . $lang . "', 'global', 'autologomod', ''),
 		('" . $lang . "', 'global', 'metaTagsOgp', '1'),
- ('" . $lang . "', 'global', 'disable_site_content', 'For technical reasons Web site temporary not available. we are very sorry for any inconvenience!')";
+ 		('" . $lang . "', 'global', 'disable_site_content', 'For technical reasons Web site temporary not available. we are very sorry for any inconvenience!')";
 
 	$sql_create_table[] = "INSERT INTO " . $db_config['prefix'] . "_setup_language (lang, setup) VALUES('" . $lang . "', 1)";
