+++ modules/contact/action.php
@@ -17,7 +17,7 @@ $sql_drop_module[] = "DROP TABLE IF EXISTS " . $db_config['prefix'] . "_" . $lan
 $sql_create_module = $sql_drop_module;
 
 $sql_create_module[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_" . $module_data . "_rows (
- id mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
+ id smallint(5) unsigned NOT NULL AUTO_INCREMENT,
  full_name varchar(255) NOT NULL,
  phone varchar(255) NOT NULL,
  fax varchar(255) NOT NULL,
@@ -31,7 +31,7 @@ $sql_create_module[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_
 
 $sql_create_module[] = "CREATE TABLE " . $db_config['prefix'] . "_" . $lang . "_" . $module_data . "_send (
  id mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
- cid mediumint(8) unsigned NOT NULL DEFAULT '0',
+ cid smallint(5) unsigned NOT NULL DEFAULT '0',
  title varchar(255) NOT NULL,
  content mediumtext NOT NULL,
  send_time int(11) unsigned NOT NULL DEFAULT '0',
